connected 4 games to main menu and removed projects for them
