+++ Connect4/connect4.h
@@ -1,0 +1,7 @@
+#pragma once
+namespace Connect4NS {
+	class Connect4 {
+	public:
+		Connect4();
+	};
+}
