+++ HangmanCode/Hangman.h
@@ -1,0 +1,8 @@
+#pragma once
+namespace HangmanNS {
+	class Hangman {
+	public:
+		Hangman();
+
+	};
+}
