--- HangmanCode/Hangman.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30204.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Hangman", "Hangman\Hangman.vcxproj", "{31C97964-ECC1-4FCA-9579-8F98EC1823AD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Debug|x64.ActiveCfg = Debug|x64
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Debug|x64.Build.0 = Debug|x64
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Debug|x86.ActiveCfg = Debug|Win32
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Debug|x86.Build.0 = Debug|Win32
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Release|x64.ActiveCfg = Release|x64
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Release|x64.Build.0 = Release|x64
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Release|x86.ActiveCfg = Release|Win32
-		{31C97964-ECC1-4FCA-9579-8F98EC1823AD}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {49AAF434-91C9-409B-B760-700D18F23EF2}
-	EndGlobalSection
-EndGlobal
