--- HangmanCode/Hangman/Hangman.h
@@ -1,6 +1,0 @@
-#pragma once
-class Hangman {
-public:
-	Hangman();
-
-};
