--- TicTacToeCode/Tic Tac Toe/Tic Tac Toe.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33213.308
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tic Tac Toe", "Tic Tac Toe\Tic Tac Toe.vcxproj", "{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Debug|x64.ActiveCfg = Debug|x64
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Debug|x64.Build.0 = Debug|x64
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Debug|x86.ActiveCfg = Debug|Win32
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Debug|x86.Build.0 = Debug|Win32
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Release|x64.ActiveCfg = Release|x64
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Release|x64.Build.0 = Release|x64
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Release|x86.ActiveCfg = Release|Win32
-		{4DC1FEC4-BD89-488C-93C6-3DC8ABB88AF3}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {833B8C8F-8A5F-49FB-A78B-3DF961BDB018}
-	EndGlobalSection
-EndGlobal
