+++ TicTacToeCode/ttt.h
@@ -1,0 +1,5 @@
+#pragma once
+class TTT {
+public:
+	TTT();
+};
