+++ WordleCode/GameManager.h
