+++ WordleCode/Wordle.h
@@ -1,0 +1,5 @@
+#pragma once
+class Wordle {
+public:
+	Wordle();
+};
