--- WordleCode/Wordle/Wordle.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33213.308
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Wordle", "Wordle\Wordle.vcxproj", "{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Debug|x64.ActiveCfg = Debug|x64
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Debug|x64.Build.0 = Debug|x64
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Debug|x86.ActiveCfg = Debug|Win32
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Debug|x86.Build.0 = Debug|Win32
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Release|x64.ActiveCfg = Release|x64
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Release|x64.Build.0 = Release|x64
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Release|x86.ActiveCfg = Release|Win32
-		{4D9F9078-BB38-4C7B-A4C3-A55539DA81AD}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C9C1EFE9-C042-41C4-B195-179E378E1BA1}
-	EndGlobalSection
-EndGlobal
