+++ WordleCode/WordleGameManager.cpp
