+++ connect4/GameManager.h
@@ -1,36 +1,37 @@
 #pragma once
+namespace Connect4NS {
+	const int COLUMNS = 7;
+	const int ROWS = 6;
+	const int GRID_SIZE = COLUMNS * ROWS;
 
-const int COLUMNS = 7;
-const int ROWS = 6;
-const int GRID_SIZE = COLUMNS * ROWS;
+	const enum SlotFills
+	{
+		NONE = 0,
+		RED,
+		YELLOW
+	};
 
-const enum SlotFills
-{
-	NONE=0,
-	RED,	
-	YELLOW
-};
+	class GameManager
+	{
+	public:
+		GameManager();
+		~GameManager() {};
 
-class GameManager
-{
-public:
-	GameManager();
-	~GameManager() {};
-	
-	void placePiece(int column);
-	void checkWin();
+		void placePiece(int column);
+		void checkWin();
 
-	int gameBoard[ROWS][COLUMNS];
-	bool isRedTurn;
-	//bool keepPlaying;
+		int gameBoard[ROWS][COLUMNS];
+		bool isRedTurn;
+		//bool keepPlaying;
 
-	// accessors and mutators
-	int getWinner();
-	int getTurnNumber();
-	void setWinner(int winner);
-	void setTurnNumber(int turnNumber);
+		// accessors and mutators
+		int getWinner();
+		int getTurnNumber();
+		void setWinner(int winner);
+		void setTurnNumber(int turnNumber);
 
-private:
-	int mTurnNumber;
-	int mWinner;
-};
+	private:
+		int mTurnNumber;
+		int mWinner;
+	};
+}
