--- connect4/connect4.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31624.102
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "connect4", "connect4.vcxproj", "{4287DEC8-3A54-4B5C-AD88-E119990E01E0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Debug|x64.ActiveCfg = Debug|x64
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Debug|x64.Build.0 = Debug|x64
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Debug|x86.ActiveCfg = Debug|Win32
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Debug|x86.Build.0 = Debug|Win32
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Release|x64.ActiveCfg = Release|x64
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Release|x64.Build.0 = Release|x64
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Release|x86.ActiveCfg = Release|Win32
-		{4287DEC8-3A54-4B5C-AD88-E119990E01E0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6CB33719-8A96-49DF-9C51-59CA46FE1C36}
-	EndGlobalSection
-EndGlobal
